requirements.txt
altgraph==0.17.4
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
exceptiongroup==1.2.0
h11==0.14.0
idna==3.5
MouseInfo==0.1.3
mss==9.0.1
outcome==1.3.0.post0
packaging==23.2
pefile==2023.2.7
Pillow==10.1.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytweening==1.0.7
pywin32-ctypes==0.2.2
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
